% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exif.R
\name{getexif}
\alias{getexif}
\title{Get Exif information}
\usage{
getexif(file, exiftool = "D:/remiPatin/Recherche/project/exiftool.exe",
  opts = NULL, intern = TRUE, simplify = FALSE)
}
\arguments{
\item{file:}{the file to be updated}

\item{exiftool:}{the path to the ExifTool binary}

\item{opts:}{additional arguments to ExifTool (optional)}

\item{intern:}{should a named vector of metadata be returned? (bool)}

\item{simplify:}{if intern==TRUE, should the results be returned as a named
vector (TRUE) or as a data.frame (FALSE)?}
}
\value{
the pdf metadata
}
\description{
\code{getexif} get pdf information from files.
}
\examples{
getexif(file, exiftool='D:/remiPatin/Recherche/project/exiftool.exe', opts=NULL, intern=TRUE, simplify=FALSE)
}

