% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfbookmarks.R
\docType{package}
\name{pdfbookmarks}
\alias{pdfbookmarks}
\alias{pdfbookmarks-package}
\title{pdfbookmarks:}
\description{
pdfbookmarks provides tools for adding bookmarks to pdf files.
adapted from https://www.donarmstrong.com/posts/adding_toc_to_pdfs_in_R/
It also provides tools for adding descriptors to pdf files.
}

