% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bookmarks.R
\name{save.bookmark}
\alias{save.bookmark}
\title{Save Bookmarks}
\usage{
save.bookmark(text, bookmarks = list(), level = 1, page = NULL)
}
\arguments{
\item{text}{the data.frame with the different variable}

\item{bookmarks}{names of the variables on which statistics are calculated}

\item{level}{names of the variable with which states are ordered}

\item{page}{type of model either 'hmm' or 'picard'}
}
\value{
a bookmark list
}
\description{
\code{save.bookmark} save bookmarks.
}
\examples{
save.bookmark(text,bookmarks=list(),level=1,page=NULL)
}

