% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_exif.R
\name{setexif}
\alias{setexif}
\title{Get Exif information}
\usage{
setexif(file, metadata,
  exiftool = "D:/remiPatin/Recherche/project/exiftool.exe")
}
\arguments{
\item{file:}{the file to be updated}

\item{metadata:}{a named character vector or list containing metadata}

\item{exiftool:}{the path to the ExifTool binary}
}
\value{
write down the new pdf file.
}
\description{
\code{getexif} get pdf information from files.
}
\examples{
setexif(file, metadata, exiftool='D:/remiPatin/Recherche/project/exiftool.exe')
}

